grace-cli-client/requirements.txt
# Grace CLI Client Requirements
# Beautiful and functional CLI application dependencies

# Core CLI libraries for beautiful interface
rich>=13.7.0
textual>=0.45.0
click>=8.1.7
typer>=0.9.0

# Screenshot and window management
pyautogui>=0.9.54
mss>=9.0.1
PyWinCtl>=0.0.50
pygetwindow>=0.0.9

# Image processing
Pillow>=10.1.0
numpy>=1.24.3

# HTTP requests for Azure OCR
requests>=2.31.0

# Data handling
pandas>=2.1.4

# Configuration management
python-dotenv>=1.0.0

# Cross-platform compatibility
psutil>=5.9.6

# Windows-specific (optional)
dxcam>=0.0.5; sys_platform == "win32"
win32gui; sys_platform == "win32"
win32con; sys_platform == "win32"

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1